refactor: move the pool and lane queries out of dnd
`dnd/` had collected three kinds of code: the pointer handlers, the
placement actions the keyboard drives, and plain queries with no drag in
them. The queries are what non-dnd code kept reaching in for — the aiming
hook and the quick-add.

`utils/pools.ts` now holds them, next to the `isPool` / `getPoolParent`
predicates they read like: `getPoolsInOrder` and `findDropPool` (the pool
to step from and the pool to aim from), `findDropSwimlane` (its lane
counterpart) and `findFreeSpotBeside`, which brings the `Direction` type
along as its only user. `use-target-aim` imports nothing from `dnd/` any
more.

`placePoolAt` becomes module-private: it has no caller outside `dnd/pools`.

The placement actions stay where they are — `dropPoolAt` shares the sizing
and placing internals with the pointer drop, and moving it would export
three helpers to save one import.
