+++ bpmn-editor/react/src/utils/index.ts
@@ -4,4 +4,5 @@ export * from './events';
 export * from './links';
 export * from './icon-contrast';
 export * from './navigation';
+export * from './pools';
 export * from './import';
